Add interaction help

+++ checkout-microservice/src/main/java/nl/tudelft/sem/template/checkout/controllers/OrderController.java
@@ -1,8 +1,10 @@
 package nl.tudelft.sem.template.checkout.controllers;
 
 import java.time.LocalDateTime;
+import java.util.Arrays;
 import nl.tudelft.sem.template.checkout.domain.Order;
 import nl.tudelft.sem.template.checkout.domain.OrderService;
+import nl.tudelft.sem.template.commons.CartPizzaAttributeConverter;
 import nl.tudelft.sem.template.commons.entity.StoreTimeCoupons;
 import nl.tudelft.sem.template.authentication.AuthManager;
 import nl.tudelft.sem.template.authentication.annotations.role.RoleRegionalManager;
@@ -32,38 +34,48 @@ public class OrderController {
         this.orderService = orderService;
     }
 
-    public long getStoreId(String storeName) {
-        String storeIdLong = requestHelper.postRequest(8084, "/store/getStoreIdFromName", storeName, String.class);
+    public long getStoreId(String storeName) throws Exception {
+        String storeIdLong = requestHelper.postRequest(8089, "/store/getStoreIdFromName", storeName, String.class);
         long storeId = Long.parseLong(storeIdLong);
         if (storeId == -1) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "This is not a real store");
         return storeId;
     }
 
     public List<CartPizza> getPizzas() {
-        return requestHelper.getRequest(8082, "/cart/getCart/" + authManager.getNetId(), List.class);
+        CartPizza[] pizzas = requestHelper.getRequest(8082, "/cart/getCart/" + authManager.getNetId(), CartPizza[].class);
+        return Arrays.asList(pizzas);
     }
 
 
     @PostMapping("/add")
     public ResponseEntity<String> addOrder(@RequestBody StoreTimeCoupons storeTimeCoupons) {
-        long storeId = getStoreId(storeTimeCoupons.getStoreName());
-        List<CartPizza> pizzas = getPizzas();
+        long storeId;
+        try {
+            storeId = getStoreId(storeTimeCoupons.getStoreName());
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body("some error");
+        }
         LocalDateTime pickupTime = storeTimeCoupons.getPickupTime();
+        if(pickupTime.minusMinutes(30).isBefore(LocalDateTime.now()))
+            return ResponseEntity.badRequest().body("Pickup time should be at least 30 minutes from order placement");
+
+        List<CartPizza> pizzas = getPizzas();
         String customer = authManager.getNetId();
 
         List<String> couponCodes = storeTimeCoupons.getCoupons();
         List<Double> pizzaPrices = getPriceForEachPizza(pizzas);
         // String finalCoupon = requestHelper.postRequest(8085, "/selectCoupon", OBJECT WITH PIZZAS AND COUPONS, String.class);
         //TODO: Wait for coupon interactions so we can retrieve the final coupon used by the user
-//        Order order = Order.builder()
-//                .withStoreId(storeId)
-//                    .withCustomerId(customer)
-//                        .withPickupTime(pickupTime)
-//                            .withPizzaList(pizzas)
-//                                .withCoupon(finalCoupon)
-//                                    .build();
+        Order order = Order.builder()
+                .withStoreId(storeId)
+                    .withCustomerId(customer)
+                        .withPickupTime(pickupTime)
+                            .withPizzaList(pizzas)
+                                //.withCoupon(finalCoupon)
+                                    .build();
+        orderService.addOrder(order);
         //TODO: Also wait for customer interaction so I can send used coupon
-        requestHelper.postRequest(8084, "store/notify", storeId, String.class); // notify store of new order
+        requestHelper.postRequest(8089, "store/notify", storeId, String.class); // notify store of new order
         return ResponseEntity.ok("Order added");
     }
 
@@ -76,9 +88,10 @@ public class OrderController {
             Order orderToBeRemoved = orderService.getOrderById(orderId);
             if (role.equals("ROLE_STORE_OWNER"))
                 return ResponseEntity.badRequest().body("Store owners can't cancel orders");
-            if (role.equals("ROLE_REGIONAL_MANGER") || orderService.getOrdersForCustomer(netId).contains(orderToBeRemoved)) {
+            if (role.equals("ROLE_REGIONAL_MANGER") ||
+                (orderService.getOrdersForCustomer(netId).contains(orderToBeRemoved) && !orderToBeRemoved.getPickupTime().minusMinutes(30).isBefore(LocalDateTime.now()))) {
                 orderService.removeOrderById(orderId);
-                return ResponseEntity.ok("Order removed");
+                return ResponseEntity.ok("Order with id " + orderId + " successfully removed");
             } else
                 return ResponseEntity.badRequest().body("Order does not belong to customer, so they cannot cancel it");
         } catch (Exception e) {
